requirements.txt
django
djangorestframework
pylint
